+++ ChangeLog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [0.0.4] - unreleased
+
+### Added
+
+- Filesystem::unlink()
+- Filesystem::chown()
+- Filesystem::chmod()
+- Filesystem::utime()
+
 ## [0.0.3] - 2024-10-25
 
 ### Added
@@ -24,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 This was the first release of fuse2rs.
 
+[0.0.4]: https://github.com/realchonk/fuse2rs/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/realchonk/fuse2rs/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/realchonk/fuse2rs/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/realchonk/fuse2rs/releases/tag/0.0.1